mv v1 logic to legacy package
